refactor(reaction_engine): split reaction_engine.rs into reaction_engine/ module directory
Split 2,800-line reaction_engine.rs into per-concern submodules so each
submodule owns its own #[cfg(test)] mod tests and compiles independently:

  reaction_engine/mod.rs        — ReactionEngine struct + dispatch entry points + re-exports
  reaction_engine/resolve.rs    — status_to_reaction_key + resolve_reaction_config + helpers
  reaction_engine/actions.rs    — dispatch_send_to_agent + dispatch_notify + dispatch_auto_merge
  reaction_engine/escalation.rs — TrackerState + parse_duration + warn_once_parse_failure

Public surface of ao_core::reaction_engine::* is unchanged. Test count: 847.
